refactor(Problem7): RecommendedFriend 클래스 분리
